+++ CHANGES.md
@@ -8,6 +8,7 @@
   - ➕ `String#clear`
   - ➕ `String#codepoints`
   - ➕ `String#deletePrefix` / `String#delete_prefix`
+  - ➕ `String#deleteSuffix` / `String#delete_suffix`
 
 ## 0.0.2 (2023-05-10)